Refactored kinematics.py into submodules.
