+++ movement/kinematics/__init__.py
@@ -5,11 +5,24 @@ from .motion import (
     compute_time_derivative,
     compute_speed,
 )
-
+from .navigation import (
+    compute_forward_vector,
+    compute_head_direction_vector,
+    compute_forward_vector_angle,
+    compute_path_length,
+)
+from .spatial import (
+    _cdist,
+    compute_pairwise_distances,
+)
 __all__ = [
     "compute_displacement",
     "compute_velocity",
     "compute_acceleration",
     "compute_time_derivative",
     "compute_speed",
+    "compute_forward_vector",
+    "compute_head_direction_vector",
+    "compute_forward_vector_angle",
+    "compute_path_length",
 ]
